+++ containers/daap-landing-to-raw/src/var/task/landing_to_raw.py
@@ -6,7 +6,7 @@ from data_platform_logging import DataPlatformLogger, s3_security_opts
 from data_platform_paths import BucketPath, RawDataExtraction, get_raw_data_bucket
 from dataengineeringutils3.s3 import read_json_from_s3
 from infer_glue_schema import infer_glue_schema_from_raw_csv
-from validation import DataInvalid, validate_data_against_schema
+from validation import DataInvalid, validate_csv_format, validate_data_against_schema
 
 s3 = boto3.client("s3")
 
@@ -22,6 +22,12 @@ def extract_columns_from_schema(schema: dict) -> dict[str, str]:
         }
     except KeyError:
         raise ValueError(f"Invalid schema: {schema}")
+
+
+def validate_format(bucket_name: str, file_key: str, logger: DataPlatformLogger):
+    obj = boto3.resource("s3").Object(bucket_name, file_key)
+    bytes_stream = obj.get()["Body"]
+    validate_csv_format(bytes_stream, logger)
 
 
 def handler(event, context):
@@ -66,10 +72,14 @@ def handler(event, context):
         inferred_columns = extract_columns_from_schema(inferred_schema)
 
         try:
+            logger.info("Validating schema")
             validate_data_against_schema(
                 inferred_columns=inferred_columns,
                 registered_schema_columns=registered_schema_columns,
             )
+            logger.info("Checking for unprocessable CSV rows")
+            validate_format(bucket_name=bucket_name, file_key=file_key, logger=logger)
+            logger.info("Continuing with ingestion")
         except DataInvalid:
             logger.error(f"{file_key} invalid; moving to fail location", exc_info=True)
             s3.copy(