+++ terraform/github/data.tf
@@ -1,0 +1,21 @@
+data "aws_caller_identity" "session" {
+  provider = aws.session
+}
+
+data "aws_iam_session_context" "session" {
+  provider = aws.session
+
+  arn = data.aws_caller_identity.session.arn
+}
+
+data "aws_secretsmanager_secret" "github_token" {
+  name = "github-token"
+}
+
+data "aws_secretsmanager_secret_version" "github_token" {
+  secret_id = data.aws_secretsmanager_secret.github_token.id
+}
+
+data "github_team" "data_and_analytics_engineering" {
+  slug = "data-and-analytics-engineering"
+}