+++ terraform/github/locals.tf
@@ -1,0 +1,3 @@
+locals {
+  users_with_special_github_access = ["julialawrence"]
+}