+++ terraform/pagerduty/terraform.tf
@@ -10,11 +10,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.23.1"
+      version = "5.24.0"
     }
     pagerduty = {
       source  = "pagerduty/pagerduty"
-      version = "3.1.0"
+      version = "3.1.1"
     }
   }
 }