+++ containers/daap-athena-load/CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - load handler attempts to load existing schema for creation of the curated table,
   falling back to inferred metadata if none is found
 
+### Changed
+
+- Updated log level from error to info when table is first created
+
 ## [1.2.3] 2023-10-18
 
 ### Changed